Lection 21: Introduces a new repository for the configuration service.
A new repository was created, not as a submodule, to contain the data of
the configuration service.
Since the new repository contains the properties defined in
application.yaml the contents of the file are no longer needed. In fact,
they would override the values of the configuration service if they
remained.